+++ CHANGELOG.md
@@ -1,3 +1,25 @@
+# CHANGELOG
+
 ## 0.0.1
 
-* TODO: Describe initial release.
+* initial release.
+
+## 0.0.2
+
+* Got Contcts with permission handler package
+
+## 0.0.3
+
+* Implemented permission handling in Method Channel
+
+## 0.0.4
+
+* Added test for Plugin and method channel
+
+## 0.0.5
+
+* Added Integration testing
+
+## 0.0.6
+
+* Changes in permission logic